improved state handling
